+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * [#225](https://github.com/tim-vandecasteele/grape-swagger/pull/225): Fixed `param_type` to have it read from parameter's documentation hash - [@zsxking](https://github.com/zsxking).
 * [#235](https://github.com/tim-vandecasteele/grape-swagger/pull/235): Fixed nested entity names in parameters and as `$ref` in models - [@frodrigo](https://github.com/frodrigo).
 * [#206](https://github.com/tim-vandecasteele/grape-swagger/pull/206): Fixed 'is_array' in the return entity being ignored - [@igormoochnick](https://github.com/igormoochnick).
+* [#245](https://github.com/tim-vandecasteele/grape-swagger/pull/245): Fixed requiring format: :json with add_swagger_documentation for grape-swagger-rails.
 
 ### 0.10.1 (March 11, 2015)